// whatsapp-clone-ui/src/app/services/models/message-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MediaContentResponse } from '../models/media-content-response';
import { RepliedMessage } from '../models/replied-message';
export interface MessageResponse {
  content?: string;
  createdAt?: string;
  forwarded?: boolean;
  id?: number;
  mediaList?: Array<MediaContentResponse>;
  repliedMessage?: RepliedMessage;
  senderId?: string;
  state?: 'SENT' | 'RECEIVED' | 'SEEN';
  type?: 'TEXT' | 'MEDIA' | 'AUDIO';
}